+++ src/main/resources/templates/usr/member/myComments.html
@@ -8,9 +8,9 @@
 <main layout:fragment="main">
     <div class="hero-content flex flex-col mb-10 2 w-[32rem] mt-5">
         <div class="flex justify-center mb-5 gap-1">
-            <a href="/usr/buyer/myPosts" class="text-2xl text-center">내 질문</a>
+            <a href="/usr/member/myPosts" class="text-2xl text-center">내 질문</a>
             <div class="divider divider-horizontal !mx-1"></div>
-            <a href="/usr/buyer/myComments" class="text-2xl font-bold text-center">내 답변</a>
+            <a href="/usr/member/myComments" class="text-2xl font-bold text-center">내 답변</a>
         </div>
         <div th:if="${comments.size()==0}" class="mt-10">
             <span class="flex items-center justify-center">등록한 답변 없습니다.</span>
