+++ src/main/resources/templates/usr/member/myPage.html
@@ -56,7 +56,7 @@
                         <svg aria-hidden="true" class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" clip-rule="evenodd"></path></svg>
                     </a>
                 </div>
-                <a href="/usr/buyer/myPosts"
+                <a href="/usr/member/myPosts"
                    class="flex items-center justify-between w-full p-5 text-left border border-gray-200 focus:ring-4 focus:ring-gray-200  hover:bg-gray-100 ">
                     <span>Q&A 게시판 활동 내역</span>
                     <svg aria-hidden="true" class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" clip-rule="evenodd"></path></svg>
